# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple crypto exchange API example.",
        "title": "Crypto Exchange API Example",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1/",
    "paths": {
        "/": {
            "get": {
                "description": "Checks the current status of the API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "API status",
                "responses": {
                    "200": {
                        "description": "Success response indicating API is OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/currencies": {
            "get": {
                "description": "Retrieves a list of all available currency pairs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "List of currency pairs",
                "responses": {
                    "200": {
                        "description": "Success response with list of currency pairs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/currencies/convert": {
            "post": {
                "description": "Converts currency from one type to another using the specified amount.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Convert currency",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Currency code to convert from",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Currency code to convert to",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Amount to convert",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with conversion result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error response when currency pair is not found or amount is invalid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/currencies/{pair}": {
            "get": {
                "description": "Retrieves detailed information of a specific currency pair.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency"
                ],
                "summary": "Currency pair details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Currency pair code",
                        "name": "pair",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with currency pair details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Error response when currency pair is not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}